move abstract worked to worker.py
